requirements.txt
flask==2.0.1
paddlepaddle==2.3.0
paddleocr==2.3.0
pillow==8.4.0
pdf2image==1.16.0
-e .